==> analysis/DEGDemo/analysis/1.prepare_gene_metadata_from_GTF.ipynb <==
# Prepare Gene Metadata

## Content
- GTF store all kinds of gene annotation information in a special format, especially the last column of GTF file contain lots of information in a string.
- Here we want to create a plain table, to extract all informations for each gene, and save it into a gene metadata

## Important pandas functions
- pd.read_csv
- pd.DataFrame.to_csv
- pd.DataFrame.apply
- pd.concat

## Load GTF

import pandas as pd

# change to your gtf location
gtf_path = '../../../data/ref/GENCODEvM24/gencode.vM24.annotation.gtf.gz'

# gtf table has some spetial formats, we need to set more parameters when read it in,
# here I just provide you the answer, you can search pandas.read_csv() documentation for more information
gtf = pd.read_csv(
    gtf_path,
    comment='#', # skip lines start with #
    sep='\t', # gtf col is sep by \t
    header=None, # no header, we will provide col names manually
    names=[
        'chrom', 'source', 'feature', 'start', 'end', 
        'na1', 'strand', 'na2', 'annotation'
    ])
gtf.head()

# this will show you documentation of the function, but search google for more examples!
?pd.read_csv

## Explore GTF content

# gtf contain different type of features
gtf['feature'].value_counts()

# here we only need those gene rows, so we do a filter here
gene_gtf = gtf[gtf['feature'] == 'gene'].copy().reset_index(drop=True)
print(f'the gene_gtf has {gene_gtf.shape[0]} rows')
gene_gtf.head()

# the last column contain lots of information in a string
example_annotation = gene_gtf['annotation'][0]
example_annotation

## Extract gene_id from the annotation column

Before prepare a whole gene metadata table, let's first do something simple to start

# we need to extract the gene_id from this string. here is how I do this with a function

def extract_gene_id(annotation):
    kv_pairs = annotation.split(';')  # split into key-value pairs
    for kv_pair in kv_pairs:  # iterate the key-value pairs
        kv_pair = kv_pair.strip(' ')  # strip removed the space
        if kv_pair.startswith('gene_id'):
            _, gene_id = kv_pair.split(' ')
            gene_id = gene_id.strip('"')  # strip removed the ""
            break
    return gene_id

extract_gene_id(example_annotation)

# now we need to apply this function on to each row of the transcript_gtf
gene_ids = gene_gtf['annotation'].apply(extract_gene_id)
# it take ~250ms in my computer
# search pandas.Dataframe.apply see more about this function
gene_ids

?pd.DataFrame.apply

## Extract everything from the annotation column and create metadata

# the annotation string is basically key-value pairs, 
# we want to turn it into a dict, then create pandas series use the dict

# how to split everything into a dict
anno_dict = {}
for pair in example_annotation.strip(';').split(';'):  # strip and split are common string methods
    k, v = pair.strip().split(' ')
    anno_dict[k] = v
anno_dict

# this function take anno str, return pd.Series
def anno_to_dict(anno_str):
    anno_dict = {}
    for pair in anno_str.strip(';').split(';'):
        k, v = pair.strip().split(' ')
        anno_dict[k] = v.strip('"')
    return pd.Series(anno_dict)

anno_to_dict(example_annotation)

# now we can apply it into every row in the gtf
gene_anno_df = gene_gtf['annotation'].apply(anno_to_dict)
gene_anno_df.head()

# human table can be different from this

# last step, we concatenate the gene_gtf with the gene_anno_df
gene_flat_table = pd.concat([gene_gtf, gene_anno_df], axis=1)

# delete some redundent or not used cols
del gene_flat_table['annotation']
del gene_flat_table['na1']
del gene_flat_table['na2']

gene_flat_table.head()

?pd.concat

# and we set gene_id as the index, which is unique and more meaningful
gene_flat_table.set_index('gene_id', inplace=True)  # inplace must be true, why? Homework
gene_flat_table.head()

## Save gene metadata

gene_flat_table.to_csv('gene_metadata.csv.gz')

?pd.DataFrame.to_csv
