==> analysis/DEGDemo/analysis/3.Pairwise_DEG_by_DESeq2_master.ipynb <==
# DEG analysis using DESeq2 Master

## Parameterize your notebook
Here I show you how to use papermill package to execute notebook template automatically with different parameter

```shell
conda install papermill
```

See papermill doc: https://papermill.readthedocs.io/en/latest/

## Input Output
- This notebook take the '3.Pairwise_DEG_by_DESeq2_template.ipynb' as template, run pairwise DEG on each pair
- The output notebook are renamed by pair, and executed automatically, together with the DEG results tables

from papermill import execute_notebook
import itertools
import pathlib

dev_times = ['E10.5', 'E14.5', 'P0']

pairs = list(itertools.combinations(dev_times, 2))
pairs

for pair in pairs:
    print('Run', pair)
    if pathlib.Path(f'{pair[0]}_vs_{pair[1]}.deg_results.csv').exists():
        print('Skip. DEG table already exists.')
        continue
    execute_notebook(
        input_path='3.Pairwise_DEG_by_DESeq2_template.ipynb',
        output_path=f'3.Pairwise_DEG_by_DESeq2.{pair[0]}_vs_{pair[1]}.ipynb',
        parameters={'sample_pair': list(pair)},
        nest_asyncio=True
    )
