==> analysis/DEGDemo/analysis/3.Pairwise_DEG_by_DESeq2_template.ipynb <==
# DEG analysis using DESeq2


## Suggested reading
- Deseq2 tutorial video part 1: https://www.youtube.com/watch?v=UFB993xufUU&t=38s
- Deseq2 tutorial video part 2: https://www.youtube.com/watch?v=Gi0JdrxRq5s
- [optional] DESeq2 documentation: http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html

## Things I do here
1. I load R packages using rpy2, make sure you installed those packages in R
2. I load gene_count_df and make a sample_df, both needed for DESeq2
3. I selected a pair of sample to run DESeq2, and the result is saved into a CSV file

## Note
- This notebook is also a demo of how we can use python and R in the same notebook
- This notebook can be run on other sample pairs, all you need to change is the parameter cell below. Try to do this with papermill automatically (either use command line interface of papermill or use papermill's `execute_notebook` function in a separate notebook)
- I usually call this parameterized notebook "template" and use it in another "master" notebook with the `execute_notebook` functions

import pandas as pd

import rpy2.robjects as ro
from rpy2.robjects.packages import importr

# this line is important, its called ipython magic, 
# and rpy2.ipython has spetial magic "%%R" that allows me to execute R directly later in this python notebook!
%load_ext rpy2.ipython

!which R

**You need to install DESeq2 in R. Make sure you are using same R version as the one printed above.**

```R
if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")

BiocManager::install("DESeq2")
BiocManager::install("apeglm")
BiocManager::install("IHW")
```

# import R packages using ryp2
importr('DESeq2')
importr('apeglm')
importr('IHW')

## Parameter

This is the key parameter cell of this notebook, you can parameterize this notebook with papermill

# here is the default value, will be replaced if this notebook is executed by papermill. Why?
sample_pair = ['E10.5', 'E14.5']

## Load data and prepare

gene_count_df = pd.read_csv('./gene_raw_count_table.csv.gz', index_col=0)
gene_count_df.head()

### Basic filter to remove low exp gene

If a gene expression is too low, it is not meaningful to report that. DESeq2 will filter genes automatically, but prefiltering will make it run faster, as suggested by the DESeq2 doc.

sample_mean_cutoff = 1

print('Before', gene_count_df.shape[0])

gene_count_df = gene_count_df[gene_count_df.mean(axis=1) > sample_mean_cutoff].copy()

print('After', gene_count_df.shape[0])

### make sample metadata based on gene_count_df.columns

sample_meta = {}
for sample in gene_count_df.columns:
    _, time, rep = sample.split('_')
    sample_meta[sample] = {'devTime': time, 'rep': rep}
sample_df = pd.DataFrame(sample_meta).T.reindex(gene_count_df.columns) # make sure order

sample_df

## Select Pair

time1, time2 = sample_pair

sample_to_use = sample_df[sample_df['devTime'].isin(sample_pair)]
gene_count_to_use = gene_count_df[sample_to_use.index]

sample_to_use

gene_count_to_use.head()

## Run DESeq2, this is very simple R code!

the DESeq2 code is very simple, but it does a lot of work behind! See the above suggested reading for more details

%%R -i gene_count_to_use -i sample_to_use -i time1 -i time2 -o deg -o counts
# this "%%R" line must be the first line of this cell, because it tells jupyter this is a R cell
# -i means import python variable into R
# python and R environment is separate! E.g., dds object will not be avaliable outside this cell, because its in R
# search rpy2 ipython magic if you want to learn more


dds <- DESeqDataSetFromMatrix(countData = gene_count_to_use,
                              colData = sample_to_use,
                              design = ~ rep + devTime)
dds <- DESeq(dds)

resultsNames(dds)

res <- results(dds, contrast=c('devTime', time1, time2), filterFun=ihw)

# get the DEG results as a dataframe
deg <- as.data.frame(res)
# get the normalized counts as a dataframe
counts <- as.data.frame(counts(dds, normalized=TRUE))

deg.head()

counts.head()

# concatenate these two tables and save it into a file
deg_with_counts = pd.concat([deg, counts], axis=1)
deg_with_counts.to_csv(f'{time1}_vs_{time2}.deg_results.csv.gz')

deg_with_counts
