==> analysis/DEGDemo/analysis/5.Prepare GSEA input files.ipynb <==
# Prepare GSEA

## Content
- GSEA needs two files as input, one is expression dataset in TXT format, the other is phenotype dataset in CLS format
- See the explaination of TXT and CLS format in GSEA doc: http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats
- This is just some spetial format needed by GSEA, so we prepare based on their requirment...

**NOTE: GSEA is mainly designed for human data, so I will not perform this with mouse data... But one the files are prepared, its pretty straight forward to run it...**

import pandas as pd
import pathlib

# make a sub dir called GSEA for new files
pathlib.Path('GSEA').mkdir(exist_ok=True)

## Load data

gene_meta = pd.read_csv('gene_metadata.csv.gz', index_col='gene_id')

## Prepare GSEA files

deg_result_paths = list(pathlib.Path().glob('*vs*.deg_results.csv.gz'))

for path in deg_result_paths:
    pair_name = '.'.join(path.name.split('.')[:-3])
    """
    Prepare expression data
    """
    
    # load the original DESeq2 output to get the normalized counts
    deg_with_norm_count = pd.read_csv(path, index_col=0)
    
    # take the last four columns, which are nrom counts
    nrom_count_df = deg_with_norm_count.iloc[:, -4:].copy()
    
    # change the index into gene names
    nrom_count_df.index = nrom_count_df.index.map(gene_meta['gene_name'])
    
    # Add modifications to match the TXT format as requied by GSEA
    # See here: http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29
    nrom_count_df.index.name = 'NAME'
    nrom_count_df['DESCRIPTION'] = 'na' # add a DESCRIPTION col
    
    # move DESCRIPTION col into the first
    col_names = list(nrom_count_df.columns)
    reordered_col_names = ['DESCRIPTION'] + col_names[:-1]
    nrom_count_df = nrom_count_df[reordered_col_names]
    
    nrom_count_df.to_csv(f'GSEA/{pair_name}.expression_data.txt', sep='\t')
    
    """
    Prepare phenotype data
    """
    # prepare the CLS format as required by GSEA
    # see here: http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29
    
    # get sample names and their class number (0 for time 1, 1 for time 2)
    time1, time2 = pair_name.split('_vs_')
    sample_names = nrom_count_df.columns[1:]
    sample_dev_times = sample_names.str.split('_').str[1]

    cls_format_str = f"""
4 2 1
# {time1} {time2}
{' '.join(sample_dev_times)}
"""
    # the above cls_format_str create a string like this:
    """    
4 2 1
# E10.5 E14.5
E10.5 E14.5 E14.5 E10.5
"""
    with open(f'GSEA/{pair_name}.phenotype_data.cls', 'w') as f:
        f.write(cls_format_str)


print(cls_format_str)

nrom_count_df
